Initial work in the update of the lwr_hw package
