+++ lwr_hw/CMakeLists.txt
@@ -23,7 +23,7 @@ find_package(catkin REQUIRED COMPONENTS
 # Depend on system install of Gazebo 
 find_package(gazebo REQUIRED)
 
-add_definitions (-fpermissive)
+add_definitions (-fpermissive -std=c++11)
 
 catkin_package(
   CATKIN_DEPENDS
@@ -31,7 +31,7 @@ catkin_package(
     controller_manager
     pluginlib
     kdl_parser
-#  INCLUDE_DIRS include
+  INCLUDE_DIRS include
   DEPENDS gazebo
 )
 
@@ -46,74 +46,44 @@ include_directories(include
 )
 link_directories(${GAZEBO_LIBRARY_DIRS})
 
-## LWR HW BASE CLASS
-add_library(lwr_hw
+## LWR HW BASE CLASS TO RULE'EM ALL
+add_library(${PROJECT_NAME}
   src/lwr_hw.cpp
+)
+target_link_libraries(${PROJECT_NAME}
+  ${catkin_LIBRARIES}
 )
 
 ## REAL 
-## lwr hw real library
+## FRI library
 add_library(friremote
   src/fri/friremote.cpp
   src//fri/friudp.cpp
 )
-add_library(lwr_hw_real
-  src/lwr_hw_real.cpp
-)
-add_dependencies(lwr_hw_real
-  lwr_hw
-  friremote
-)
-target_link_libraries(lwr_hw_real
-  lwr_hw
-  friremote
-  ${catkin_LIBRARIES}
-)
-
 ## lwr hw real node
-add_executable(lwr_hw_node
-  src/lwr_hw_real_node.cpp
+add_executable(lwr_hw_node  src/lwr_hw_real_node.cpp
+  include/${PROJECT_NAME}/lwr_hw_real.hpp
+)
+add_dependencies(lwr_hw_node
+  friremote
+  ${PROJECT_NAME} 
 )
 target_link_libraries(lwr_hw_node
-  lwr_hw_real
+  friremote
+  ${PROJECT_NAME}
   ${catkin_LIBRARIES}
 )
 
 ## SIMULATION
-
-## lwr hw sim base class
-# add_library(lwr_hw_sim
-#  src/lwr_hw_sim.cpp
-#)
-
-## lwr hw sim default behavior
-add_library(default_lwr_hw_sim 
-  src/default_lwr_hw_sim.cpp
-)
-# add_dependencies(default_lwr_hw_sim
-#  lwr_hw_sim
-#)
-target_link_libraries(default_lwr_hw_sim
-#  lwr_hw_sim
-  lwr_hw # this is needed because there is no lwr_hw_sim library
-  ${catkin_LIBRARIES} 
-  ${GAZEBO_LIBRARIES}
-)
-
 # lwr hw sim gazebo plugin
-add_library(lwr_hw_sim_plugin
-  src/lwr_hw_sim_plugin.cpp
+add_library(lwr_hw_sim_plugin  src/lwr_hw_sim_plugin.cpp
+  include/${PROJECT_NAME}/lwr_hw_sim.hpp
+)
+add_dependencies(lwr_hw_sim_plugin
+  ${PROJECT_NAME}
 )
 target_link_libraries(lwr_hw_sim_plugin
+  ${PROJECT_NAME}
   ${catkin_LIBRARIES}
   ${GAZEBO_LIBRARIES}
 )
-
-## Install
-install(TARGETS default_lwr_hw_sim
-  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  )
-
-install(FILES lwr_hw_sim_plugins.xml
-  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
-  )
