--- lwr_hw/config/joint_names.yaml
@@ -1,8 +1,0 @@
-joints:
-  - lwr_0_joint
-  - lwr_1_joint
-  - lwr_2_joint
-  - lwr_3_joint
-  - lwr_4_joint
-  - lwr_5_joint
-  - lwr_6_joint
