+++ lwr_hw/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>lwr_hw</name>
-  <version>0.1.0</version>
+  <version>0.2.0</version>
   <description>The lwr_hw package implementes real and simulation interfaces to the LWR 4+</description>
 
   <maintainer email="cjrosales@gmail.com">Carlos J. Rosales</maintainer>
@@ -35,7 +35,6 @@
   <run_depend>joint_limits_interface</run_depend>
 
   <export>
-    <lwr_hw plugin="${prefix}/lwr_hw_sim_plugins.xml"/>
     <gazebo_ros plugin_path="${prefix}/lib"/>
   </export>
 </package>
