+++ lwr_hw/src/lwr_hw_real_node.cpp
@@ -4,13 +4,14 @@
 #include <time.h>
 #include <signal.h>
 #include <stdexcept>
+#include <boost/shared_ptr.hpp>
 
 // ROS headers
 #include <ros/ros.h>
 #include <controller_manager/controller_manager.h>
 
 // the lwr hw real interface
-#include "lwr_hw/lwr_hw_real.h"
+#include "lwr_hw/lwr_hw_real.hpp"
 
 bool g_quit = false;
 
@@ -33,24 +34,33 @@ int main( int argc, char** argv )
   signal(SIGINT, quitRequested);
   signal(SIGHUP, quitRequested);
 
-  // construct the lwr
+  // create a node
   ros::NodeHandle lwr_nh("");
-  lwr_hw::LWRHWreal lwr_robot(lwr_nh);
 
-  // configuration routines
-  lwr_robot.start();
+  // get params or give default values
+  int port;
+  std::string hintToRemoteHost;
+  lwr_nh.param("port", port, 49939);
+  lwr_nh.param("ip", hintToRemoteHost, std::string("192.168.0.10") );
+
+  // construct and start the real lwr
+  lwr_hw::LWRHWreal lwr_robot;
+  lwr_robot.create();
+  lwr_robot.setPort(port);
+  lwr_robot.setIP(hintToRemoteHost);
+  if(!lwr_robot.init())
+  {
+    ROS_FATAL_NAMED("lwr_hw","Could not initialize robot real interface");
+    return -1;
+  }
 
   // timer variables
   struct timespec ts = {0, 0};
   ros::Time last(ts.tv_sec, ts.tv_nsec), now(ts.tv_sec, ts.tv_nsec);
   ros::Duration period(1.0);
 
-  //realtime_tools::RealtimePublisher<std_msgs::Duration> publisher(lwr_nh, "loop_rate", 2);
-
   //the controller manager
   controller_manager::ControllerManager manager(&lwr_robot, lwr_nh);
-
-  //uint32_t count = 0;
 
   // run as fast as possible
   while( !g_quit ) 
@@ -70,30 +80,14 @@ int main( int argc, char** argv )
     } 
 
     // read the state from the lwr
-    if(!lwr_robot.read(now, period))
-    {
-      g_quit = true;
-      break;
-    }
+    lwr_robot.read(now, period);
 
     // update the controllers
     manager.update(now, period);
 
     // write the command to the lwr
     lwr_robot.write(now, period);
-
-    // if(count++ > 1000)
-    // {
-    //   if(publisher.trylock())
-    //   {
-    //     count = 0;
-    //     publisher.msg_.data = period;
-    //     publisher.unlockAndPublish();
-    //   }
-    // }
   }
-
-  //publisher.stop();
 
   std::cerr<<"Stopping spinner..."<<std::endl;
   spinner.stop();
