+++ single_lwr_example/single_lwr_robot/config/lwr_hw.yaml
@@ -1,0 +1,3 @@
+lwr:
+  root: lwr_base_link
+  tip: lwr_7_link
